Give partial SCSS source files leading underscores
SCSS understand the idea of "partial" source files -- those that are
intended only for inclusion in another file using the `@import`
directive. By naming partials with a leading underscore, the distinction
between "partial" and "entrypoint" CSS files is made clear.

This has no impact on the renderered CSS files, which are byte-for-byte
identical after this commit.