Clean up definition of mixins/vars/etc.
This commit adds an '_all.scss' mixin module, which imports all mixins
(none of which should output any CSS) in an appropriate order.

It then renames the 'variables' partial to 'base', and adds an import at
the bottom which loads all mixins.

This pattern allows us to easily and clearly define stylesheets by
simply '@import'ing the base partial at the start, ensuring that all
mixins and variables are available.