Simplify CSS imports
Adding a "base partial" which defines variables and includes all mixins
makes tracking allowable imports in CSS files much simpler.

This commit removes unused/redundant imports from partial CSS files so
that the following now applies:

- entrypoint CSS files *only* import partials (no mixins)
- no partials other than the base partial import anything

In addition, this commit renames '_common.scss' to '_help-page.scss' --
it was previously styles "common" to the web application and the client,
but now the help page is the only page with these styles.