Stop passing request into RenameUserService
The reason to have the request was that we want to commit the
transaction before indexing, and then the `BatchIndexer` needs the
request (for sending the `AnnotationTransformEvent`). This commit adds a
new parameter to the service for passing in a `reindex` function which
takes care of this. Now there is no dependency on the request anymore
for this service.