Add ability to configure aggregations for search queries
Works exactly the same as matchers and filters. Agreggations are classes
defined in `h.api.search.query` and can be added to a query with
`append_aggregation`. These aggregation classes need to provide a
`parse_result` method which will get the Elasticsearch output and is
expected to parse the data into something more manageable for callers of
the search module. These results will then be part of the returned
`SearchResult` instance.