Return a SearchResult object when searching (#3622)
Now that we don't just blindly render what ElasticSearch returns we can
change the return value of `h.api.search.core.search` to be a value
object for easier access and sane default values. This is also done as a
preparation for aggregation support to search.

We can now also make Elasticsearch not render the search result
documents as we're going to return a list of ids any way.