Make HTML enhancements URL-independent
This implements a new system for upgrading elements on the page with
additional functionality which does not require hardcoding the routes
that might use such functionality into the client-side code.

Instead a map of selector to controller classes is created in the root
site bundles and the `upgradeElements` module handles upgrading all
matching elements on the page.

This also gives us a central place to put logic relating to upgrades to
catch issues such as errors upgrading elements or elements which should
be upgraded but for which no matching controller was found.