Move config of auth/authz policies into h.auth/authz modules
Rather than relying on each app instance to manually call
`set_authentication_policy` on the configurator, this commit moves the
(default) authentication policy configuration into the `h.auth` and
`h.authz` modules.

This will make it easier for the exact configuration of the
authentication policy to depend on application settings.