Move configuration of auth policies into h/auth/__init__.py
This is a slight reversion to some changes made earlier, as it makes
reconfiguring the auth policies in response to application settings
easier.

This commit changes h.auth.policy.AuthenticationPolicy to be an generic
policy which accepts two child authentication policies to be used for
the API and as a fallback.

The WebSocketAuthenticationPolicy, a thin wrapper around a
MultiAuthenticationPolicy, is removed, in favour of simply
constructing the appropriate policy in h/auth/__init__.py.