Enable telling devserver which processes to run (#3641)
Add options to the `hypothesis devserver` command to tell it which of
the processes you would like it to run and which not, instead of it
always running all the processes.

The problem with `hypothesis devserver` is that it doesn't work with pdb
(or any similar Python debugger). Because devserver multiplexes the
output of multiple processes into one terminal, while you were trying to
interact with pdb in say the web process output from the celery process
would be printed and ruin your pdb shell. Also, devserver doesn't know
which of its multiple processes to send the input you're typing to, so
you can't interact with pdb.

Now, if you want to use pdb, you can run `hypothesis devserver --no-web`
and then run the web process separately in another shell (by running the
`gunicorn` command directly) and use pdb as normal.

Or you could debug the celery worker process in the same way but with
`hypothesis devserver --no-worker`.