Implement new nav bar for Activity Pages and Accounts pages
This implements the static parts of the new site navigation bar.  The
dynamic parts will be added in subsequent commits.

The icons for the search bar are created with inline SVG, generated
using a helper in the web service. This is similar to the approach
GitHub describe at
https://github.com/blog/2112-delivering-octicons-with-svg

In order for icons to be colorable using CSS, they must set the 'fill'
or 'stroke' properties to 'currentColor'

The SVG icon helper uses `lxml.etree` instead of `xml.etree` to benefit
from support for namespace prefixes and avoid the problem described at
http://stackoverflow.com/questions/8983041