Enable submitting accounts forms via XHR
Enable submitting the change email and change password forms via
XMLHttpRequest, receiving back just the <form> element as an HTML
snippet instead of the whole HTML page.

Also move the .render() calls for the email and password forms into the
whole page template, so that they're only rendered when the whole page
is rendered, i.e. only for non-XHR requests. (For XHR requests, the new
to_xhr_response() renders just the form that was submitted.)