Update Docker build for new packaging
The updated packaging of h and memex allow us to simplify the Docker
build substantially. In this commit we switch back to using a `docker
build` based on the output of `git archive` rather than using the source
distribution tarball -- the source distribution tarball is now for memex
and not the whole of h.

This has the substantial advantage that we no longer invalidate the
Docker layer caching of all of our dependencies on every commit, making
most Docker builds (especially those that do not modify requirements or
the memex version number) much faster.