Remove build dependencies at end of application build
Until now, every commit invalidated the Docker image layer which
contained the application dependencies. This meant that installing the
system-level build deps and leaving them installed was preferable,
despite the additional space they occupied, as a result of the decreased
build time.

Now, we no longer need to invalidate the layer on every commit, so it
makes more sense to recover the disk space by only installing the system
build deps for as long as they are needed.

This reduces the size of the image by ~100MB.

We now also run the tests against the installed source in the container,
rather than (ab)using the Jenkins workspace. This will ensure that we
don't accidentally depend on data from a previous test run.