Run each alembic migration in its own transaction
This is a config setting that we should have enabled a long time ago
(and the alembic default should be turned on).
It ensures that each migration is wrapped in its own transaction. This
is especially important when we run two migrations at the same time, the
first one adding a column, and the second one using that column.

We also don't need to wrap the `context.run_migrations()` call in a
transaction anymore since alembic takes care of that now.