Add custom return values to handle_form_submission()
Allow the on_success() callback to return a value that
handle_form_submission() will then return instead of its default
HTTPFound(location=request.url) if form validation is successful and the
request is not an XHR.

If on_success() returns None then handle_form_submission() will still
return its default HTTPFound(location=request.url).

This is needed because not all form submissions want to redirect to the
same URL that the form was submitted from. For example the create group
form creates a new group and then redirects to the new group's URL.