Add the "authority" column to the user model and set on signup
Having now added the column directly to the database, it is safe to add
reference to the authority column to the user model.

We ensure that the authority column is filled in for new users using the
`default_authority` parameter to the `UserSignupService`, which defaults
to `request.auth_domain`.

This means that in production, new users will be created with an
authority of "hypothes.is" as expected.