Extract paginate()
Extract the core logic for generating the data structure for the paginator
control's template into a paginate() function.

paginate_query() now contains just the code specific to paginating
sqlalchemy queries, and paginate() can be re-used by further
paginate_*() functions for paginating other kinds of objects.

Change the tests to test the new core paginate() function directly
instead of testing paginate_query().

Also split up these tests a little: they still use parametrize() but
there's a separate parametrized test for each of the values ('cur',
'next', 'prev', 'max') that paginate() returns. This adds a little more
test code duplication but I think it makes the tests more descriptive
and the test code easier to understand.

The paginate_query() tests are then rewritten to use a mock paginate()
to just test that paginate_query() uses paginate(), the view callable,
and the query that the view callable returns correctly. These tests are
also refactored a little to make use of the mock library and pytest
fixtures, which makes the tests nice and short and simple.