Add user/group activity pages and auto-redirects
This commit adds user and group activity pages at

     /groups/{pubid}/search
     /users/{username}/search

which automatically filter the searches on these pages by the relevant
group or user. These are not necessarily final URLs for these pages.

It also adds functionality to the main activity page route at /search so
that if a query is entered containing a single user or group term, then
the user is redirected to the relevant user or group activity page, with
their query appropriately modified.

For example, a request to

     /search?q=group:abcd1234+tag:foo

will be automatically redirected to

     /groups/abcd1234/search?q=tag:foo

If a group is specified in the query to a group search page, such as

     /groups/abcd1234/search?q=group:efgh4567

then it will simply be ignored. Displaying the union of the results for
two groups on a specialised group search page doesn't make a lot of
sense.