Use UserService to retrieve user in authenticated_user
We aim soon to remove `h.accounts.get_user` in favour of the user
service.

I've taken the liberty of changing this function from using
`request.current_route_url()` to plain old `request.url` as it makes it
a little easier to test. It shouldn't affect the functionality at all.