Make password fields nullable
There's no reason we need to require a password when creating a user in
the database -- the User#check_password method can deal with this
scenario and ensure that such a user can never log in.

Making these fields nullable means that we can create users in tests
without needing to set passwords, which is (necessarily) a slow
operation.