Don't set passwords when not necessary in tests
Setting a password on a user is necessarily a slow operation. Doing this
many times for test users substantially slows down the running of the
test suite.

This commit removes all the places where we're setting passwords for
users when we don't need to.

In my testing this results in a ~1.7× speedup for our current test
suite.

Timings for three runs of `tox -e py27-h` before this change:

     36s
     36s
     36s

Timings for three runs of `tox -e py27-h` after this change:

     20s
     21s
     22s

Speedup = 36s/21s ≈ 1.7×