Add a passlib CryptContext for managing passwords
There are a couple of problems with our current password storage -- none
of these are obviously security problems, but they are a bit ugly and it
would be nice to fix them.

This commit is part of a couple of changes which will replace
cryptacular -- a perfectly serviceable library but one which hasn't been
updated in 4 years and has no support for newer version of the bcrypt
algorithm -- with passlib[1] and bcrypt[2]. In particular, it adds a
CryptContext to the `h.security` package which will be used to manage
the hashing and validation of user passwords.

This gives us a couple of things:

- We are no longer (for new passwords) "double-salting" them --
  eventually we should be able to remove the user.salt column from the
  database, as the password hashes include an automatically generated
  salt.

- The salt now changes whenever the password changes. This happened
  before, but it wasn't obvious that this was the case, because of the
  confusion around double-salting: only the cryptacular-generated salt
  changed, but User.salt did not.

- We can now easily change the number of rounds used by bcrypt, or even
  the entire hashing algorithm, simply by modifying the password context
  in h.security. No changes to the model will be required.

[1]: https://pythonhosted.org/passlib/
[2]: https://github.com/pyca/bcrypt/