Fix and simplify style bundle rebuilds when styles change
There were a couple of mistakes in the previous logic which caused
rebuilds not to happen reliably when style files were changed:

 1. The list of files imported by a bundle was not returned
    from compileSass() after the addition of PostCSS into the
    pipeline, so changes to imported files did not trigger a rebuild

 2. If a build of a newly added file failed, the newly added dependency
    was not added to the watched file set and hence when the file was
    fixed, no rebuild was triggered.

This commit reverts to just using a `gulp.watch` and rebuilding all
bundles when any style file changes or is added. This is conceptually
not as nice as rebuilding a file only when its dependencies change ala.
watchify, but its simpler for the moment.