Cache users fetched by the UserService
Because we're fetching users by their userid, and not by their primary
key, we can't take advantage of the SQLAlchemy identity map to avoid
fetching the same user more than once within a single session.

This commit adds our own identity map to the UserService, in order to
ensure that we don't need to fetch the same user more than once within a
single request.