Strip POST payload data from Sentry reports
As useful as it can be, sending POST payloads to Sentry will inevitably
result in us sending unhashed user passwords to a third-party service.
Not a great idea.

The SanitizePasswordsProcessor is already on by default. This commit
adds the RemovePostDataProcessor, which strips the POST payload data
from error reports, as SanitizePasswordsProcessor is not smart enough to
handle multipart forms.