Add blacklist for group names
Since the URL structure is /{pubid}/{slug}, and we have two special URLs
(/{pubid}/leave and /{pubid}/edit) we have to stop allowing group names
that end up with slugs that are "edit" or "leave".

This implements a blacklist similar to what we do for usernames, the
error message says that the chosen group name is not allowed.

Fixes #3775