Add facility to override detected browser environment
Add support for overriding the environment classes that are set on the
`<html>` element by using `__env__=$flags` in the URL, where $flags is a
semi-colon separated list of flags that indicate flags to add or remove
(if the flag starts with a "no-" prefix).

For example:

Simulate touch input:

  https://hypothes.is/signup?__env__=touch

Simulate JS load timeout:

  https://hypothes.is/signup?__env__=js-timeout

Simulate failure to run JS:

  https://hypothes.is/signup?__env__=no-js-capable
  https://hypothes.is/signup?nojs=1

This replaces the earlier `nojs=1` mechanism in `upgradeElements` which
only prevented controllers from being initialized but did not alter the
environment classes applied to `<html>`.