Clear the zope.sqlalchemy session state cache before creating a session
We appear to be fighting a consistency bug in zope.sqlalchemy, in which
some request transactions don't clear their session statuses out of
zope.sqlalchemy.datamanager._SESSION_STATE. When this happens, entries
relating to old and long-dead sessions accumulate in _SESSION_STATE
until the memory address of an old session is reused for a new one.
(CPython only guarantees the uniqueness of id(object) for objects with
overlapping lifetimes.)

An ugly but effective hack is to manually clear the _SESSION_STATE
memo whenever we create a new request, which happens to be safe for us
because we only ever have a single session open within the context of a
single request.