Paginate search results properly by page number
Even though we already set the `page` parameter, it didn't get passed
down to the search query. To do that we also have to convert the page
number to a pair of limit/offset. This can all be done transparently in
the `query` module, while the page size definition should stay in the
view because it's being used for the pagination as well as for the
query.