Do not open a database session for asset/debug requests
Opening a database session to load feature flags for these requests is
not only not necessary, it is quite problematic. It causes the leakage
of references to the database session inside `zope.sqlalchemy` because
these requests have been opted out of the transaction manager by an
`activate_hook` (see `h.app`).

By using the database in these requests, we were accidentally
registering a database session with a transaction manager that wasn't
hooked up to the request lifecycle. This meant that zope.sqlalchemy's
internal references to the session weren't properly cleaned up. In a
high traffic environment -- these leaked session references would then
later lead to a session (in an entirely different request) not being
correctly registered with the transaction manager.

I've also opened a pull request against `pyramid_tm` which should
prevent this class of bug in the future:

    https://github.com/pylons/pyramid_tm/pull/46