Add character limit counter to <input>s
Add a character limit counter JavaScript enhancement to Deform <input> form
fields, the same as we have for <textarea>s, whenever the <input> has a
max_length argument to its Deform widget in its Deform schema.

Currently this causes the group name field to have a character counter
when creating or editing a group, since it's the only <input> that has
currently has a max_length.