Improve search bucket behavior in the JS load timeout case
When a search result bucket is upgraded after a JS load timeout, we want
the results to remain initially expanded, but they should still toggle
when clicked. Previously the toggle functionality was broken after an
eventual load.

This commit fixes the problem by removing the static `env-js-timeout` styling
and instead passing the timeout state to the search bucket controller,
which can use it to set the initial expanded state of the results.

 * Fix search bucket toggling failing after JS load timeout

 * Fix search bucket background being white instead of grey when
   expanded if mouse was not hovered over bucket.

   The `is-expanded` state was being set on the wrong element in
   `SearchBucketController#update`

 * Make successful initialization of `window.envFlags` a requirement
   for upgrades to be applied. In very rare cases, it may happen that
   header.js fails to load but site.js does load, in that case fall
   back to the plain HTML page.