Increase JavaScript tooltip hover target size
Show the tooltip when the mouse hovers anywhere over the <label>, i.e.
over the entire "ORCID (i)", rather than just over the (i).

This is more accessible (larger hover target is easier to hit) and also
more consistent with normal HTML title-attribute tooltips which are
usually attached to words such as <label> or <a> elements and not (just)
to icons.

You no longer need to hover over the (i) icon to reveal the tooltip, but
the icon still serves the purpose of giving the user a hint that there
is a tooltip here.