Retry mailer tasks on SMTP failure
We are seeing evidence that occasionally the Mandrill SMTP server
disconnects our mailer, or perhaps that the network between our
application and Mandrill occasionally drops:

  https://sentry.io/hypothesis/prod/issues/152186124/

This commit adds retry logic to the mailer task to ensure that in the
event of any kind of network or SMTP error, the task is requeued for
another delivery attempt later.

With the current configuration we will attempt to deliver mail four
times (i.e. 3 retries) over a period of approximately 20 minutes, using
an exponential backoff.