Don't make hint icon touch target larger on mobile
Remove the code that was making the form field "hint" icon's touch
target larger on mobile devices. You can now touch anywhere in the
entire "Field Name (i)" label, rather than having to touch the (i) icon,
so the touch target is already large enough. Additionally the resizing
of the touch target on mobile was breaking the alignment of the tooltip
with the icon, and simply removing the resizing was the easiest way to
fix this.