Don't crash when there are no matching annotations
The bucketing code expects to receive a list of annotations, not a
sqlalchemy query object. Sending it a query object crashes when there
are 0 annotations that match the query, because the bucketing code's
test for this relies on an empty list being falsey (whereas a sqlalchemy
query with no results is truthy).

This could happen for example if the ?page=n argument in the query
string is larger than the number of pages of matching annotations.

Fixes #3842