Add an AuthClient model for publisher accounts
An AuthClient represents an OAuth client, an entity which can access
protected resources (such as annotations) on behalf of a user.

The first type of OAuth client we have is a very special one, which can
access protected resources for *any* user within its *authority*. These
are "publisher" accounts, which can create users in our database, and
subsequently issue grant authorization tokens for any of those users.