Add schema migrations to update user uniqueness constraints
Add two migrations to update the uniqueness constraints on the user
table.

The first, b0e1a12de5e8, renames the existing constraints, and
then adds new constraints in a way that won't interfere with table
writes. This is a reversible migration

The second, 94c989e06363, removes the old single-column constraints that
were renamed by the first migration.