Don't strip whitespace from non-title document metas
We don't allow None, empty strings, whitespace-only strings, or strings
with leading or trailing whitespace, as items in the `value` list of a
document meta whose `type` is 'title'. Nor do we alluw empty `value`
lists. But we _do_ want to allow all of these kinds of values for
non-'title' document metas (at least for now).