Replace NipsaUser with nipsa column on User model
This commit updates the NipsaService and the admin NIPSA views to use
the newly-added "nipsa" column on the User table rather than the
independent NipsaUser table.

This ensures that users remain flagged if we change their username, and
removes the need for an additional table in the database and additional
code in the admin views.