Fix a bug where MixedCase usernames weren't correctly NIPSA'd
Previously, trying to flag a user with a mixed-case username (e.g.
"DonkeyFace") would result in that user's past annotations not being
retroactively hidden.

This commit ensures that the various queries that are used to find and
update the user's annotations correctly coerce the userid to lowercase
before using them in a term filter.

In future, we should perhaps not have our own custom NIPSA queries, but
have a generic "reindex this user's annotations" Celery job.