Don't crash on null document.highwire.pdf_url
A POST like this to the create or update annotation API would cause
parse_document_claims to crash:

    http POST 'http://localhost:5000/api/annotations'
      Authorization:'Bearer 6879-***'
      uri='http://example.com'
      document:='{"highwire": {"pdf_url": null}}'

And this would cause a different crash:

    http POST 'http://localhost:5000/api/annotations'
      Authorization:'Bearer 6879-***'
      uri='http://example.com'
      document:='{"highwire": {"pdf_url": [null]}}'

The problem is that parse_document_claims assumes that
annotation.document.highwire.pdf_url is a list of strings.

Fix both crashes by having AnnotationSchema reject the field if
it's not a list of strings, so it never even gets to
parse_document_claims. (This is the same way that we handle other
document metadata fields: AnnotationSchema makes sure they're the right
type, parse_document_claims can then safely assume this.)