Use new ticket policy and migrate session auth states
We first try and get the authenticated userid from the new token policy,
if that returns `None` we try and get it from the old session policy.
When the session policy returns a userid, we then remember that userid
in the new fallback policy. This way we get to migrate all active
sessions within 7 days, authentication states that didn't get migrated
within this time are not something we need to care about, as these users
would get logged out anyway.