Rename .title to .title_or_filename_or_uri
DocumentHTMLPresenter's .title property returns the document's title and
falls back to its filename or URI if it has no title.

This is not the behaviour that we want for the new activity search page
- we just want the title or "Untitled document".

But I can't just change what .title does because other parts of the code
are using it.

So start by renaming it, paving the way for adding multiple
.title_or_* properties to DocumentHTMLPresenter.