Add client helper for performing XHR submissions
Add a `submitForm()` function which handles submitting a form on the
client and returning a promise that resolves to the re-rendered form and
status code, or the error if form submission fails.

The modern `fetch()` API is used rather than XHR since this provides a
more convenient Promise-based interface. It is polyfilled for older
browsers.