Fix search icon alignment
Fix the vertical position of the looking-glass icon in the search field
on the /search page.

The height of the <input> that contains this icon (including padding and
margin) is 23px. The height of the icon itself is 16px. That leaves
23 - 16 = 7px to be evenly distributed above and below the icon. So
there should be 7/2 = 3.5px between the top of the icon and the top of
the <input>s border.

If you actually enter 3.5px into the CSS then the icon looks slightly
too high to my eye. Perhaps Chrome rounds down to 3px. Imo it looks
better (less noticeable) to have the thin (and diagonal) handle of the
looking glass hanging slightly too far down, than to have its large
circle too high up. So I went with 4px.