Centralise route definitions in h/routes.py
Our route naming (and even our URL selection) has at times in the past
been a little inconsistent and messy. Partly in order to fix this, and
partly in order to enable us to move all views into h.views, this commit
centralises the definition of the application's routes into a single
file.

Hopefully this makes it easier to tell where a URL comes from and easier
to ensure consistent naming styles across the application in future.