Don't wrap URLs unnecessarily
Don't unnecessarily wrap URLs shown inside document buckets - expand the
div that the URL is wrapped within to fill all of the horizontal space
available to it (but leaving a 10px gap to the right, same size gap as
used to the left and right of document buckets throughout).

Fixes #3741.