Fix ragged document bucket title left-alignment
Flexbox was variably growing and shrinking the width taken up by the
domain name depending on the width's of the domain name and title
contents, resulting in ragged left-alignment of document bucket titles,
especially on smaller screens.

Tell flexbox that the width of the domain name area is fixed, so that
the document bucket titles are always left-aligned with each other as
the flex box grows and shrinks.