Remove an unused CSS media query
This media query used to make it so that, below a certain screen size,
the domain name of a document bucket would be layed out above the title
(and everything else).

The HTML restructuring that was done in ec8c98593ea0e4276d4b714949e4356cfc8e5c3e
means that this media query no longer doed the desired layout change. It
only adds 10px of vertical margin (which without the layout change is
pointless).

A new media query needs to be written to achieve the same thing with the
new HTML layout. But first, the old broken media query should be
deleted.