Add a helper for extracting HTTP Basic auth credentials
This commit adds a helper to `h.auth.util` for extracting any HTTP Basic
access authentication credentials present in a request. There are a
number of pitfalls here, mostly stemming from the fact that the
interaction of HTTP Basic authentication with Unicode is very poorly
defined in standards.

This is not likely to cause us problems in practice, as we're issuing
ASCII-only client IDs and secrets.