Add a preliminary "create user" API endpoint
In order to allow third-party accounts clients to furnish their users
with grant tokens that allow them to use our API, we need those users to
first exist in our database.

This commit adds a relatively straightforward "create user" API
endpoint that checks the client's credentials (which must be provided
as an HTTP Basic username and password) and creates a user from provided
parameters if they validate.

At this stage, we haven't hooked up any useful validation of the payload
provided by the HTTP client.